feat(deploy): Allow target account to list objects in build bucket
This change updates the S3 bucket access policy to grant the target account `s3:ListBucket` permissions.
This is necessary to allow the target account to list the objects in the build bucket, so that it can
download a whole directory or to test if s3 command is working or not.