refactor(api_server): improve auth extraction logic
This commit refactors the authentication extraction logic in the API
server. The `Authentication` struct is now wrapped in a new
`AuthFromHeaders` struct, which is an `Option<Authentication>`.

This change allows the server to handle requests as normal if the auth
information is OK, even if the `allow_missing_or_bad_signature` is
being set. Only when auth information is really bad or missing, we would
like to set the auth to None.