refactor: remove rpc_io_uring
The major reason to try rpc_io_uring is for the zero copy send feature,
however, the performance numbers were really bad and it is unknown to
make it better than tokio epoll stack. Remove this feature for now since
we are tunning arena allocator which would be hard to make rpc_io_uring
work.